Add GunSlot class and added list of GunSlots to HullConfigs
