+++ main/prevShip.ini
@@ -1,3 +1,3 @@
 hull=imperialSmall
 money=20
-items=rep rep rep a1 s1 blaster sloMoCharge sloMoCharge sloMoCharge 
+items=rep rep a1 s1 blaster sloMoCharge sloMoCharge sloMoCharge
