+++ main/src/com/miloshpetrov/sol2/game/SaveManager.java
@@ -4,7 +4,7 @@ import com.miloshpetrov.sol2.files.FileManager;
 import com.miloshpetrov.sol2.files.HullConfigManager;
 import com.miloshpetrov.sol2.game.item.ItemManager;
 import com.miloshpetrov.sol2.game.item.SolItem;
-import com.miloshpetrov.sol2.game.ship.HullConfig;
+import com.miloshpetrov.sol2.game.ship.hulls.HullConfig;
 import com.miloshpetrov.sol2.IniReader;
 
 import java.util.ArrayList;
