+++ main/src/com/miloshpetrov/sol2/game/item/ShipItem.java
@@ -3,7 +3,7 @@ package com.miloshpetrov.sol2.game.item;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.miloshpetrov.sol2.game.SolGame;
-import com.miloshpetrov.sol2.game.ship.HullConfig;
+import com.miloshpetrov.sol2.game.ship.hulls.HullConfig;
 
 public class ShipItem implements SolItem {
 
@@ -19,15 +19,15 @@ public class ShipItem implements SolItem {
   public static String makeDesc(HullConfig hull) {
     StringBuilder sb = new StringBuilder();
     sb.append("Takes ").append(hull.getMaxLife()).append(" dmg\n");
-    boolean noG2 = hull.getG2Pos() == null;
-    if (noG2 || hull.m1IsFixed() != hull.m2IsFixed()) {
-      if (noG2) {
-        sb.append(hull.m1IsFixed() ? "1 heavy gun slot\n" : "1 light gun slot\n");
+    boolean secondGunSlot = hull.getNrOfGunSlots() > 1;
+    if (!secondGunSlot || hull.getGunSlot(0).allowsRotation() != hull.getGunSlot(1).allowsRotation()) {
+      if (!secondGunSlot) {
+        sb.append(!hull.getGunSlot(0).allowsRotation() ? "1 heavy gun slot\n" : "1 light gun slot\n");
       } else {
         sb.append("1 heavy + 1 light gun slots\n");
       }
     } else {
-      sb.append(hull.m1IsFixed() ? "2 heavy gun slots\n" : "2 light gun slots\n");
+      sb.append(!hull.getGunSlot(0).allowsRotation() ? "2 heavy gun slots\n" : "2 light gun slots\n");
     }
     if (hull.getAbility() != null) {
       sb.append("Ability:\n");
