+++ main/src/com/miloshpetrov/sol2/game/item/TradeConfig.java
@@ -3,7 +3,7 @@ package com.miloshpetrov.sol2.game.item;
 import com.badlogic.gdx.utils.JsonValue;
 import com.miloshpetrov.sol2.files.HullConfigManager;
 import com.miloshpetrov.sol2.game.ShipConfig;
-import com.miloshpetrov.sol2.game.ship.HullConfig;
+import com.miloshpetrov.sol2.game.ship.hulls.HullConfig;
 
 import java.util.*;
